# adapter/src/main/resources/member-openapi.yaml
openapi: 3.0.3
info:
  version: "1.0.0"
  title: Member API
servers:
  - url: http://localhost:8080/
tags:
  - name: Member
    description: Actions On Members
paths:
  /member/list:
    get:
      summary: List all members
      operationId: getMemberList
      tags:
        - members
      responses:
        200:
          description: A paged array of members
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /member:
    get:
      summary: Get one specified member
      operationId: getMember
      tags:
        - members
      parameters:
        - name: id
          in: query
          required: true
          description: The ID of the member to retrieve
          schema:
            type: string
      responses:
        200:
          description: A member object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /member/create:
    post:
      summary: Create a new member
      operationId: createMember
      tags:
        - members
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberRequest'
        required: true
      responses:
        204:
          description: Member successfully created
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    MemberResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - birthDate
        - email
        - phone
        - address
      properties:
        id:
          type: string
        firstName:
          type: string
          description: The first name of the member
        lastName:
          type: string
          description: The last name of the member
        birthDate:
          type: string
          format: date
          description: The birth date of the member
        email:
          type: string
          format: email
        phone:
          type: string
          description: The phone number of the member
        address:
          type: string
          description: The address of the member
    MemberRequest:
      type: object
      required:
        - firstName
        - lastName
        - birthDate
        - email
        - phone
        - address
      properties:
        firstName:
          type: string
          description: The first name of the member
        lastName:
          type: string
          description: The last name of the member
        birthDate:
          type: string
          format: date
          description: The birth date of the member
        email:
          type: string
          format: email
          description: The email address of the member
        phone:
          type: string
          description: The phone number of the member
        address:
          type: string
          description: The address  of the member
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string